requirements.txt
asttokens==2.4.1
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==22.0.0
idna
ipython==8.23.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
packaging==24.0
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy
stack-data==0.6.3
stripe==8.8.0
traitlets==5.14.2
typing_extensions==4.10.0
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.1
